mentioning test case facilitation, metadata for recordings.

+++ app/src/main/java/com/elses/respository/BeaconMetrics.java
@@ -5,13 +5,13 @@ public class BeaconMetrics{
     public Long timestamp;
     public String name;
     public int txPower;
-    public int distance;
+    public double distance;
 
     public BeaconMetrics(){
 
     }
 
-    public BeaconMetrics(String name, int rssi,int txPower,int distance){
+    public BeaconMetrics(String name, int rssi,int txPower,double distance){
         this.rssi=rssi;
         this.txPower=txPower;
         this.distance=distance;
